+++ packages/stg-evm-v2/devtools/tasks/index.ts
@@ -74,7 +74,8 @@ import {
 } from '@stargatefinance/stg-devtools-v2'
 import { subtask, task } from 'hardhat/config'
 
-import { createConnectedContractFactory, inheritTask } from '@layerzerolabs/devtools-evm-hardhat'
+import { createConnectedContractFactory, inheritTask, types } from '@layerzerolabs/devtools-evm-hardhat'
+import { createLogger } from '@layerzerolabs/lz-utilities'
 import {
     SUBTASK_LZ_OAPP_CONFIG_LOAD,
     SUBTASK_LZ_OAPP_WIRE_CONFIGURE,
@@ -626,10 +627,25 @@ interface ConfigFile {
 
 task(TASK_STG_GET_CONFIG_HASHES, 'get config for a token')
     .addFlag('genJson', 'Generate also each config as a json file')
+    .addParam('stage', 'Chain stage. One of: mainnet, testnet, sandbox', undefined, types.stage, true)
     .setAction(async (args, hre) => {
-        const directoryPath: string = path.join(__dirname, '..', 'config', 'mainnet', '01')
-        const directoryPathJson: string = path.join(__dirname, '..', 'config', 'mainnet', '01', 'json')
-        const directoryPathJsonHashes: string = path.join(__dirname, '..', 'config', 'mainnet', '01', 'hashes')
+        let chainsConfigPath: string
+        if (args.stage === 'mainnet') {
+            chainsConfigPath = path.join(__dirname, '..', 'config', 'mainnet', '01')
+        } else if (args.stage === 'testnet') {
+            chainsConfigPath = path.join(__dirname, '..', 'config', 'testnet')
+        } else if (args.stage === 'sandbox') {
+            chainsConfigPath = path.join(__dirname, '..', 'config', 'sandbox')
+        } else {
+            throw new Error('Invalid stage')
+        }
+
+        const logger = createLogger(process.env.LOG_LEVEL || 'info')
+        logger.info(`Getting config info for ${args.stage} stage`)
+
+        const directoryPath: string = chainsConfigPath
+        const directoryPathJson: string = path.join(chainsConfigPath, 'json')
+        const directoryPathJsonHashes: string = path.join(chainsConfigPath, 'hashes')
         try {
             // Read all files in the directory
             const files = fs.readdirSync(directoryPath)
@@ -717,6 +733,7 @@ task(TASK_STG_GET_CONFIG_HASHES, 'get config for a token')
                     }
                 })
             )
+            if (args.genJson) logger.info(`Wrote config for ${args.stage} stage to ${directoryPathJson}`)
 
             // create folder if it doesn't exist
             if (!fs.existsSync(directoryPathJsonHashes)) {
@@ -727,9 +744,10 @@ task(TASK_STG_GET_CONFIG_HASHES, 'get config for a token')
                 path.join(directoryPathJsonHashes, 'hashes.json'),
                 JSON.stringify(output, null, 2)
             )
+            logger.info(`Wrote hashes for ${args.stage} stage to ${directoryPathJsonHashes}/hashes.json`)
             return output
         } catch (error) {
-            console.error('Error reading directory:', error)
+            logger.error('Error reading directory:', error)
             return []
         }
     })